第三次作业

+++ HomeWork3/HomeWork3_1/Point.cs
@@ -2,11 +2,11 @@ namespace HomeWork3_1
 {
     public class Point
     {
-        public double X { get; set; }
+        public int X { get; set; }
 
-        public double Y { get; set; }
+        public int Y { get; set; }
 
-        public Point(double x, double y)
+        public Point(int x, int y)
         {
             X = x;
             Y = y;
